refactor: rename file view
